+++ app/components/Error.tsx
@@ -1,0 +1,10 @@
+import { useInstantSearch } from 'react-instantsearch-core'
+
+export function Error() {
+  const { error } = useInstantSearch({ catchError: true })
+
+  if (error) {
+    throw error
+  }
+  return <></>
+}
